feat(api): คะแนนข้อสอบรองรับทศนิยมขั้นต่ำ 0.5
ทั้งสายคะแนนของข้อสอบเคยเป็นจำนวนเต็ม การตั้ง 0.5 จึงพังเงียบ ๆ:
MySQL ปัดเศษให้เอง (0.5 -> 1) โดยไม่มี error และ validation ก็ตีกลับ 422

- migration: แปลง 6 คอลัมน์เป็น decimal (questions.points,
  user_answer_questions.points, lesson_answer_questions.points,
  course_quizzes.total_score, course_quiz_results.score, courses.total_score)
  ใช้ raw ALTER เพื่อคุม nullability/default ให้ตรงเป๊ะ และข้ามถ้าไม่ใช่ mysql
- validation: required|numeric|min:0.5|max:1000|multiple_of:0.5
  ทั้งข้อสอบรายวิชาและคำถามของบทเรียน
- cast points/score เป็น float ใน 5 โมเดล
  CourseMember เคย cast achieved_score/bonus_points เป็น integer
  ทั้งที่คอลัมน์จริงเป็น decimal อยู่แล้ว -> ตัดเศษทิ้งทันทีที่มีคะแนน .5
- calculateInternalTotalScore/syncCourseTotalScore คืนค่า float
  (ของเดิมเป็น : int จะตัด 2.5 เหลือ 2 และ PHP 8.4 จะ throw)
- import รับ .0/.5, export เขียนเป็น float
- เทสต์ใหม่ 3 เคส: สร้างข้อ 0.5 ได้ / ตีกลับ 0.3 กับ 0 / รวมคะแนนได้ 2.5

ตรวจแล้วบน MySQL จริง: up() แปลงครบ 6 คอลัมน์ · down() คืนสคีมาเดิมตรงทุกคอลัมน์
ข้อมูลเดิมครบ (questions 2,345 แถว sum เท่าเดิม) และ 0.5 เก็บได้จริงเป็น '0.50'

+++ api/nuxnanravel/app/Models/CourseMember.php
@@ -22,8 +22,8 @@ class CourseMember extends Model
     {
         return [
             'edited_grade' => 'float',
-            'achieved_score' => 'integer',
-            'bonus_points' => 'integer',
+            'achieved_score' => 'float',
+            'bonus_points' => 'float',
             'external_score_points' => 'float',
             'efficiency' => 'integer',
             'grade_progress' => 'float',
